Restored reconnection feature
Feature broken after 448e9a2
io.sockets.connected is not a substitute for
gameState.players[id]connected for tracking connected players.
gameState.players[id].connected is set to false when a player
disconnects because we want to let the server know that we have player
data without a player. when the player reconnects or a substitute player
connects the new socket id is assigned to any available disconnected
player data allowing the new connection to take over for the previously
abandoned player data.